// src/components/ui/features/experimental/legacy-stacking-nav-motion/types.ts
export type EasingType = 
  | 'linear'
  | 'easeIn'
  | 'easeOut'
  | 'easeInOut'
  | 'circIn'
  | 'circOut'
  | 'circInOut'
  | 'backIn'
  | 'backOut'
  | 'backInOut'
  | 'anticipate'
  | 'expoIn'
  | 'expoOut'
  | 'expoInOut'
// ...
export type ButtonVariant =
  | 'primary'
  | 'secondary'
  | 'tertiary'
  | 'shine'
  | 'tab'
  | 'link-gray'
  | 'link-color'